Extract Pages.py file
