+++ minis/downloader/__init__.py
@@ -1,1 +1,2 @@
-
+from .downloadAllMinis import *
+from .Pages import *
